chore_: move extkeys submodule to cmd
