+++ account/generator/account.go
@@ -6,10 +6,11 @@ import (
 	"encoding/json"
 	"time"
 
+	"github.com/status-im/status-go/cmd/extkeys"
+
 	accountJson "github.com/status-im/status-go/account/json"
 	"github.com/status-im/status-go/eth-node/crypto"
 	"github.com/status-im/status-go/eth-node/types"
-	"github.com/status-im/status-go/extkeys"
 	"github.com/status-im/status-go/multiaccounts"
 )
