+++ account/generator/utils.go
@@ -4,9 +4,10 @@ import (
 	"bytes"
 	"errors"
 
+	"github.com/status-im/status-go/cmd/extkeys"
+
 	"github.com/status-im/status-go/eth-node/crypto"
 	"github.com/status-im/status-go/eth-node/types"
-	"github.com/status-im/status-go/extkeys"
 )
 
 var (
