+++ account/generator/utils_test.go
@@ -3,11 +3,12 @@ package generator
 import (
 	"testing"
 
+	"github.com/status-im/status-go/cmd/extkeys"
+
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/status-im/status-go/eth-node/types"
-	"github.com/status-im/status-go/extkeys"
 )
 
 func generateTestKey(t *testing.T) *extkeys.ExtendedKey {
