+++ account/onboarding.go
@@ -3,12 +3,13 @@ package account
 import (
 	"fmt"
 
+	"github.com/status-im/status-go/cmd/extkeys"
+
 	"github.com/pborman/uuid"
 
 	"github.com/status-im/status-go/account/generator"
 	"github.com/status-im/status-go/eth-node/crypto"
 	"github.com/status-im/status-go/eth-node/types"
-	"github.com/status-im/status-go/extkeys"
 )
 
 // OnboardingAccount is returned during onboarding and contains its ID and the mnemonic to re-generate the same account Info keys.
