+++ account/utils_test.go
@@ -4,13 +4,14 @@ package account
 import (
 	"testing"
 
+	"github.com/status-im/status-go/cmd/extkeys"
+
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 
 	"github.com/status-im/status-go/account/generator"
 	"github.com/status-im/status-go/eth-node/crypto"
 	"github.com/status-im/status-go/eth-node/types"
-	"github.com/status-im/status-go/extkeys"
 )
 
 type AccountUtilsTestSuite struct {
