+++ cmd/extkeys/hdkey_test.go
