+++ cmd/extkeys/mnemonic.go
