+++ cmd/extkeys/mnemonic_test.go
