+++ cmd/extkeys/mnemonic_vectors.json
