+++ cmd/extkeys/utils.go
