+++ eth-node/keystore/passphrase.go
@@ -12,13 +12,14 @@ import (
 	"encoding/json"
 	"fmt"
 
+	"github.com/status-im/status-go/cmd/extkeys"
+
 	"github.com/google/uuid"
 	"golang.org/x/crypto/pbkdf2"
 	"golang.org/x/crypto/scrypt"
 
 	"github.com/status-im/status-go/eth-node/crypto"
 	"github.com/status-im/status-go/eth-node/types"
-	"github.com/status-im/status-go/extkeys"
 )
 
 const (
